+++ lib/max_country_picker.dart
@@ -1,5 +1,2 @@
-library max_country_picker;
-
-export 'model/country_model.dart';
 export 'config/country_list_config.dart';
 export 'view/country_picker.dart';