+++ CHANGELOG.md
@@ -6,3 +6,8 @@ Initial release
 ## 0.0.1+1
 
 Fixing immutable error
+
+## 1.0.0+2
+
+Remove additional package
+Fixing widget conflic